+++ newton/_src/solvers/kamino/_src/solver_kamino_impl.py
@@ -14,7 +14,7 @@ import warp as wp
 
 # Newton imports
 from ....core.types import override
-from ....sim import Contacts
+from ....sim import Contacts, State
 from ...solver import SolverBase
 
 # Kamino imports
@@ -592,11 +592,11 @@ class SolverKaminoImpl(SolverBase):
         self._write_step_output(state_out=state_out)
 
     @override
-    def notify_model_changed(self, flags: int):
+    def notify_model_changed(self, flags: int) -> None:
         pass  # TODO: Migrate implementation when we fully integrate with Newton
 
     @override
-    def update_contacts(self, contacts: Contacts) -> None:
+    def update_contacts(self, contacts: Contacts, state: State | None = None) -> None:
         pass  # TODO: Migrate implementation when we fully integrate with Newton
 
     @override