+++ newton/_src/utils/import_mjcf.py
@@ -1003,12 +1003,11 @@ def parse_mjcf(
                 shapes.append(s)
 
             elif geom_type == "ellipsoid":
-                # MuJoCo ellipsoid size is (rx, ry, rz) semi-axes, same as Newton a, b, c
                 s = builder.add_shape_ellipsoid(
                     xform=tf,
-                    a=geom_size[0],
-                    b=geom_size[1],
-                    c=geom_size[2],
+                    rx=geom_size[0],
+                    ry=geom_size[1],
+                    rz=geom_size[2],
                     **shape_kwargs,
                 )
                 shapes.append(s)